Fallback to broker_selector_first_available for broker_selector_usage
Due to a bug somewhere else, the usage of all brokers is set to 65535 (integer max) for all brokers. The algorithm of broker_selector_usage fails, as it selects the first broker (line 1138: best = 0) if all usages are equal and does not check weather the first broker is available.

So I suggest this fallback to first_available. That algorithm works although the usage values are useless.